// public/lx-a/_reference/lexus-encore-download/account.lexus.com.au/src/types/dateTime.ts
export type Minutes = number;

export enum KeazDayNumber {
    Monday = 0,
    Tuesday = 1,
    Wednesday = 2,
    Thursday = 3,
    Friday = 4,
    Saturday = 5,
    Sunday = 6,
}

export interface Interval {
    start: Minutes;
    end: Minutes;
}

export enum JSMonthNumber {
    January = 0,
    February = 1,
    March = 2,
    April = 3,
    May = 4,
    June = 5,
    July = 6,
    August = 7,
    September = 8,
    October = 9,
    November = 10,
    December = 11,
}

export enum JSDayNumber {
    Sunday = 0,
    Monday = 1,
    Tuesday = 2,
    Wednesday = 3,
    Thursday = 4,
    Friday = 5,
    Saturday = 6,
}